+++ Sources/VCore/Extensions/Foundation/Array/Array.ReversedOnCondition.swift
@@ -12,13 +12,28 @@ extension Array {
     /// Returns reversed `Array` if condition is met.
     ///
     ///     ["London", "Paris", "New York"]
-    ///         .reversed(if: reversesOrder)
+    ///         .reversed(reversesOrder)
     ///
-    public func reversed(if condition: Bool) -> Self {
+    public func reversed(
+        _ condition: Bool
+    ) -> Self {
         if condition {
             self.reversed()
         } else {
             self
+        }
+    }
+
+    /// Reverses `Array` if condition is met.
+    ///
+    ///     var array: [String] = ["London", "Paris", "New York"]
+    ///     array.reversed(reversesOrder)
+    ///
+    mutating public func reverse(
+        _ condition: Bool
+    ) {
+        if condition {
+            reverse()
         }
     }
 }