+++ CHANGELOG.md
@@ -78,6 +78,7 @@ Extensions
 - `Binding.unwrappedBinding(default:)` method is added that wraps `Optional` `Binding` with a default value
 - `View.onTouchInteraction(perform:)` method is added that performs action when `View` detects touch down or touch up interactions
 - `DragGesture.Value.velocity(inRelationTo:)` method is removed, as it's obsoleted by native `velocity` property
+- `Array.reverse(_:)` is added to complement `Array.reversed(_:)`
 
 API